feat: Refactor Playback component and update stories to TypeScript and use correct props
